pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fastapi-supabase"
version = "0.1.0"
description = "FastAPI backend with Supabase authentication integration"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
    { name = "Bruno", email = "bwagner57@gmail.com" }
]

dependencies = [
    "fastapi>=0.100.0",
    "pyjwt>=2.8.0",
    "uvicorn>=0.22.0",
]

[project.optional-dependencies]
dev = [
    "python-dotenv>=1.0.0",
    "httpie>=3.2.1",
]

[tool.hatch.build.targets.wheel]
packages = ["src/fastapi_supabase"]

[tool.hatch.build]
include = [
    "src/fastapi_supabase/**/*.py",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/fastapi-supabase"
"Bug Tracker" = "https://github.com/yourusername/fastapi-supabase/issues"
